Add new role to combat invader cores: exterminator
The exterminator is a melee creep that will spawn in whenever a remote mining room is occupied by an invader core. Each room has an exterminator cap of two. They are statically sized to [5x ATTACK, 5x MOVE].
When both are attacking the core, it shouldn't take longer than 300 ticks to get rid of it. Once they kill the core, they will check to see if there are any other hostile creeps in the room. If not, they will
commit suicide, and remote mining activities should resume as normal.